requirements.txt
streamlit
pytesseract
numpy
pandas
pillow